+++ lib/util/epd/driver/uc8253.dart
@@ -2,7 +2,7 @@ import 'driver.dart';
 
 // UC8253 commands/registers,
 // define in the epaper display controller (UC8253) reference manual
-class Uc8252 extends Driver {
+class Uc8253 extends Driver {
   @override
   int get refresh => 0x12;
   @override
